User auth

+++ src/features/catalog/api/updateUserRating.ts
@@ -1,28 +1,31 @@
-import { supabaseClient } from '@utils/supabase-client';
+import { User } from '@supabase/gotrue-js';
+import { SupabaseClient } from '@supabase/supabase-js';
 
-async function getUser() {
-  const {
-    data: { user },
-  } = await supabaseClient.auth.getUser();
+export async function deleteUserRating(
+  nfid: number,
+  user: User,
+  supabaseClient: SupabaseClient
+) {
+  if (user) {
+    const { error } = await supabaseClient
+      .from('rating')
+      .delete()
+      .eq('user_item_key', `${user?.id}-${nfid}`);
 
-  return user;
-}
-export async function deleteUserRating(nfid: number) {
-  const user = await getUser();
-  const { error } = await supabaseClient
-    .from('rating')
-    .delete()
-    .eq('user_item_key', `${user?.id}-${nfid}`);
-
-  if (error) {
-    console.log('Error:', {
-      message: error.message,
-      details: error.details,
-    });
+    if (error) {
+      console.log('Error:', {
+        message: error.message,
+        details: error.details,
+      });
+    }
   }
 }
-export async function updateUserRating(nfid: number, stream: boolean) {
-  const user = await getUser();
+export async function updateUserRating(
+  nfid: number,
+  stream: boolean,
+  user: User,
+  supabaseClient: SupabaseClient
+) {
   if (user) {
     const { error } = await supabaseClient.from('rating').upsert({
       user_item_key: `${user.id}-${nfid}`,
